Moved all the tasks to a separate sub-package as .tasks
MatchDataParser, TeamDataParser and FootBallSyncTask
